pyproject.toml
[tool.poetry]
name = "blob-quickstart-v12"
version = "0.1.0"
description = ""
authors = ["Moein"]

[tool.poetry.dependencies]
python = "^3.10"
azure-storage-blob = "^12.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
